requirements.txt
mysqlclient
decouple
django-fontawesome
